Fold the reading protocols into AppDatabase instead of a runtime cast
MetricsProvider and ActivityProvider no longer probe the backend with an
as? cast to pick the native-aggregation path. MetricSeriesReading and
ActiveUsersReading are now part of the shared AppDatabase and BackendDatabase
surface with default implementations that return nil, so backends that cannot
aggregate natively (CloudKit) fall through to that default while the providers
call database.metricSeries and database.activeUsers directly and treat a nil
result as the cue to fall back to the matrix query. Now that both protocols are
broadly used they move back into the shared Core/Backend grouping, and the
CLAUDE.md rule recommending co-location of cast-probed protocols is dropped.
